changes to type id

+++ C3PO-back-end/src/main/java/com/codeforall/online/c3po/command/PlayerDto.java
@@ -12,7 +12,7 @@ public class PlayerDto {
     private String username;
     private int score;
 
-    private Integer id;
+    private Long id;
 
     public String getUsername() {
         return username;
@@ -26,11 +26,11 @@ public class PlayerDto {
         return score;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
